requirements.txt

async-lru==2.0.4
certifi==2024.2.2
chardet==5.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
imgurpython==1.1.7
Plate==1.0.0
pyaes
Pyrogram
PySocks
requests==2.31.0
rfc3986==2.0.0
sniffio==1.3.0
urllib3==2.2.0
XlsxWriter==3.1.9
Pillow
openpyxl
google-cloud
cloud-sql-python-connector
pymysql
sqlalchemy
pandas
openpyxl-image-loader == 1.0.5
TgCrypto
mysql
mysql-connector-python